Topic/rails4.2

+++ app/models/maintenance.rb
@@ -16,11 +16,11 @@ class Maintenance < ActiveRecord::Base
   #Allocate NPCs/create NPCs and assign to existing kingdom
   def self.new_kingdom_npcs(kingdom)
     @@report << "NEW KINGDOM NPCS FOR " + kingdom.name
-    kingdom.npcs.find(:all, :conditions => ['is_hired = ?', 0]).each{
-      |uh| uh.update_attribute(:kingdom_id, nil) if rand > 0.75 }
+    kingdom.npcs.where(is_hired: 0).each {
+        |uh| uh.update_attribute(:kingdom_id, nil) if rand > 0.75 }
 
-    @unhired_merchants = kingdom.merchants.find(:all, :conditions => ['is_hired = 0'])
-    @unhired_guards = kingdom.guards.find(:all, :conditions => ['is_hired = 0'])
+    @unhired_merchants = kingdom.merchants.where(is_hired: false)
+    @unhired_guards    = kingdom.guards.where(is_hired: false)
 
     if @unhired_merchants.size < kingdom.kingdom_empty_shops.size * 1.5
       npc_solicitation(kingdom, NpcMerchant)
@@ -29,9 +29,9 @@ class Maintenance < ActiveRecord::Base
   end
 
   def self.npc_solicitation(kingdom, npc_class)
-    if @unhired_merchants.size < kingdom.player_character.level * 2
+    if kingdom.merchants.where(is_hired: false).size < kingdom.player_character.level * 2
 
-      @npc_from_pool = npc_class.find(:first, :conditions => ['kingdom_id is NULL'], :order => 'rand()')
+      @npc_from_pool = npc_class.order('rand()').find_by(kingdom_id: nil)
 
       if @npc_from_pool.nil? || rand > 0.75
         @new_guy = npc_class.generate(kingdom.id)
@@ -52,47 +52,49 @@ class Maintenance < ActiveRecord::Base
   def self.kingdom_npcs_maintenance(kingdom, npcs)
     for npc in npcs
       #get the disease damage toll first
-      @disease_damage = npc.illnesses.inject(0){|c,i| c+i.disease.HP_per_turn }
+      @disease_damage = npc.illnesses.inject(0) { |c, i| c+i.disease.HP_per_turn.to_i }
 
       #lock NPC, just to be safe
       Npc.transaction do
         begin
-        npc.lock! && npc.stat.lock! && npc.health.lock!
+          npc.lock! && npc.stat.lock! && npc.health.lock!
 
-        Stat.symbols.each{|sym| npc.stat[sym] += rand(4) }
-        npc.gold = rand(500)
-        npc.experience += rand(50)
-        npc.health.base_HP += rand(7)
-        npc.health.HP += npc.health.base_HP - @disease_damage
+          Stat.symbols.each { |sym| npc.stat[sym] += rand(4) }
+          npc.gold           = rand(500)
+          npc.experience     += rand(50)
+          npc.health.base_HP += rand(7)
+          npc.health.HP      += npc.health.base_HP - @disease_damage
 
-        @terminal_diseases = npc.illnesses.find(:all, :include => 'disease', :conditions => ['diseases.NPC_fatal = true'])
-        if @terminal_diseases.size > 0
-          if npc.health.HP <= 0
-            npc.health.wellness = SpecialCode.get_code('wellness','dead')
-            KingdomNotice.create(  npc.name + " died from " + @terminal_diseases.rand.disease.name + ".",
-                                  kingdom.id, "Minister of Health and Sanitation")
+          @terminal_diseases = npc.illnesses.joins(:disease).where(diseases: {NPC_fatal: true})
+          if @terminal_diseases.size > 0
+            if npc.health.HP <= 0
+              npc.health.wellness = SpecialCode.get_code('wellness', 'dead')
+              KingdomNotice.create_notice(
+                  npc.name + " died from " + @terminal_diseases.rand.disease.name + ".",
+                  kingdom.id,
+                  "Minister of Health and Sanitation")
+            end
+          else
+            npc.health.HP = [npc.health.HP, 1].max
           end
-        else
-          npc.health.HP = [npc.health.HP,1].max
-        end
-        npc.save! && npc.stat.save! && npc.health.save!
+          npc.save! && npc.stat.save! && npc.health.save!
         rescue Exception => e
           @@report << e
         end
       end
 
-      if npc.health.wellness == SpecialCode.get_code('wellness','dead')
+      if npc.health.wellness == SpecialCode.get_code('wellness', 'dead')
         @@report << "This is a dead NPC :'("
       else
-        if npc.npc_merchant_detail
+        if npc.kind_of?(NpcMerchant) and npc.npc_merchant_detail
           if npc.npc_merchant_detail.healing_sales.to_i > 0
             #can any pandemics be cured?
             for pandemic in kingdom.pandemics
-              if HealerSkill.find(:first, :conditions => ['disease_id = ? AND min_sales <= ?',
-                                  pandemic.disease_id, npc.npc_merchant_detail.healing_sales]) &&
-                  rand(pandemic.disease.virility) < npc.int / 10  #then NPC has cured the pandemic
+              if HealerSkill.where(disease_id: pandemic.disease_id) \
+                     .find_by('min_sales <= ?', npc.npc_merchant_detail.healing_sales) &&
+                  rand(pandemic.disease.virility) < npc.int / 10 #then NPC has cured the pandemic
                 text = npc.name + " has discovered a cure for the " + pandemic.disease.name +
-                       " pandemic which had been tormenting the kingdom for " + pandemic.day + " days."
+                    " pandemic which had been tormenting the kingdom for " + pandemic.day + " days."
                 KingdomNotice.create_notice(text, kingdom.id, "Minister of Health and Sanitation")
                 pandemic.destroy
               end
@@ -118,8 +120,8 @@ class Maintenance < ActiveRecord::Base
       pandemic.save
 
       #only airbourne can be spread
-      if pandemic.disease.trans_method == SpecialCode.get_code('trans_method','air')
-        npcs.each{|npc| Illness.infect(npc, pandemic.disease) }
+      if pandemic.disease.trans_method == SpecialCode.get_code('trans_method', 'air')
+        npcs.each { |npc| Illness.infect(npc, pandemic.disease) }
       end
 
       Kingdom.transaction do
@@ -128,18 +130,18 @@ class Maintenance < ActiveRecord::Base
         if kingdom.num_peasants < pandemic.disease.min_peasants
           #not enough peasants, pandemic ends
           text = "The number of cases of " + pandemic.disease.name + " has greatly diminished, ending the " +
-                  pandemic.day.to_s + " pandemic."
+              pandemic.day.to_s + " pandemic."
           KingdomNotice.create_notice(text, kingdom.id, "Minister of Health and Sanitation")
           pandemic.destroy
         else
           @fatality = pandemic.disease.peasant_fatality
           @fatality = @fatality / 2.0 + rand(@fatality / 2.0)
-          @deaths = (kingdom.num_peasants * (@fatality / 100.0)).to_i
+          @deaths   = (kingdom.num_peasants * (@fatality / 100.0)).to_i
 
           @dead_from_disease += @deaths
 
           #notice for deaths
-          text = @deaths.to_s + " peasants died from " + pandemic.disease.name
+          text               = @deaths.to_s + " peasants died from " + pandemic.disease.name
           KingdomNotice.create_notice(text, kingdom.id, "Minister of Health and Sanitation")
           kingdom.num_peasants -= @deaths
         end
@@ -152,12 +154,11 @@ class Maintenance < ActiveRecord::Base
 
   #main routine to take care of all the kingdom maintenance that needs done
   def self.kingdom_maintenance
-    @kingdoms = Kingdom.find(:all, :conditions => ['id > 0'])
     @updates = 0
-    for kingdom in @kingdoms
+    Kingdom.where('id > 0').includes(pandemics: :disease).each do |kingdom|
       @@report << "Maintenance for " + kingdom.name
 
-      @npcs = kingdom.live_npcs
+      @npcs           = kingdom.live_npcs.includes(illnesses: :disease)
       @disease_deaths = kingdom_pandemics(kingdom, @npcs)
       kingdom_npcs_maintenance(kingdom, @npcs)
 
@@ -167,34 +168,34 @@ class Maintenance < ActiveRecord::Base
 
         @current_pop = kingdom.num_peasants
         @tax_revenue = (kingdom.num_peasants * kingdom.tax_rate / 1000.0).to_i
-        @immigrants = 3 + (100 - (kingdom.tax_rate ** 2) / 100.0).to_i
-        @immigrants *= [(kingdom.housing_cap - kingdom.num_peasants - @immigrants), 1].min
-        @immigrants = 0 if @immigrants < 0
-        @emmigrants = [30+kingdom.housing_cap - kingdom.num_peasants, 0].min
-        @emmigrants /= (rand(10)+1) unless @emmigrants < kingdom.housing_cap * -2
-        @births = ((rand(10) + kingdom.num_peasants + @emmigrants) * [100 - kingdom.tax_rate ** 2, 1].max / 1000.0).to_i
+        @immigrants  = 3 + (100 - (kingdom.tax_rate ** 2) / 100.0).to_i
+        @immigrants  *= [(kingdom.housing_cap - kingdom.num_peasants - @immigrants), 1].min
+        @immigrants  = 0 if @immigrants < 0
+        @emmigrants  = [30+kingdom.housing_cap - kingdom.num_peasants, 0].min
+        @emmigrants  /= (rand(10)+1) unless @emmigrants < kingdom.housing_cap * -2
+        @births      = ((rand(10) + kingdom.num_peasants + @emmigrants) * [100 - kingdom.tax_rate ** 2, 1].max / 1000.0).to_i
 
         kingdom.num_peasants += @births + @immigrants + @emmigrants
-        kingdom.gold += @tax_revenue
+        kingdom.gold         += @tax_revenue
 
         text = "Population trends:" +
-               "%-26s%20s" % ["<br/>Last Population:" ,@current_pop] +
-               "%-26s%20s" % ["<br/>Current Population:" ,kingdom.num_peasants] +
-               "%-26s%20s" % ["<br/>Births:", @births.to_s] +
-               "%-26s%20s" % ["<br/>Immigrants:", @immigrants.to_s] +
-               "%-26s%20s" % ["<br/>Emmigrants:", (-1*@emmigrants).to_s] +
-               "%-26s%20s" % ["<br/>Deaths from disease:", @disease_deaths.to_s]
+            "%-26s%20s" % ["<br/>Last Population:", @current_pop] +
+            "%-26s%20s" % ["<br/>Current Population:", kingdom.num_peasants] +
+            "%-26s%20s" % ["<br/>Births:", @births.to_s] +
+            "%-26s%20s" % ["<br/>Immigrants:", @immigrants.to_s] +
+            "%-26s%20s" % ["<br/>Emmigrants:", (-1*@emmigrants).to_s] +
+            "%-26s%20s" % ["<br/>Deaths from disease:", @disease_deaths.to_s]
         KingdomNotice.create_notice(text, kingdom.id, "Minister of Population")
 
         kingdom.save!
       end
 
       @@report << "\nPopulation trends:" +
-                  "%-26s%20s" % ["\nCurrent Population:" ,@current_pop.to_s] +
-                  "%-26s%20s" % ["\nBirths:", @births.to_s] +
-                  "%-26s%20s" % ["\nImmigrants:", @immigrants.to_s] +
-                  "%-26s%20s" % ["\nEmmigrants:", (-1*@emmigrants).to_s] +
-                  "%-26s%20s" % ["\nDeaths from disease:", @disease_deaths.to_s]
+          "%-26s%20s" % ["\nCurrent Population:", @current_pop.to_s] +
+          "%-26s%20s" % ["\nBirths:", @births.to_s] +
+          "%-26s%20s" % ["\nImmigrants:", @immigrants.to_s] +
+          "%-26s%20s" % ["\nEmmigrants:", (-1*@emmigrants).to_s] +
+          "%-26s%20s" % ["\nDeaths from disease:", @disease_deaths.to_s]
 
       #take care of the new NPC stuff, only if there is a king
       if kingdom.player_character_id
@@ -209,41 +210,44 @@ class Maintenance < ActiveRecord::Base
   #mostly just gives players characters more time units, but could be expanded later.
   def self.player_character_maintenance
     #add turns for player characters that are active only, and have used turns since last time
-    @player_characters = PlayerCharacter.find(:all, :conditions => ['char_stat = ? and turns < 200', SpecialCode.get_code('char_stat','active')])
-    @updated_count = 0
-    for pc in @player_characters
-      begin
-      #might as well get the lock
-      PlayerCharacter.transaction do
-        @@report << "Updating PC #" + pc.id.to_s + " " + pc.name
+    @player_characters =
+        @updated_count = 0
+    PlayerCharacter.where(char_stat: SpecialCode.get_code('char_stat', 'active'))
+        .where('turns < 200').find_in_batches do |pc_group|
+      pc_group.each do |pc|
+        begin
+          #might as well get the lock
+          PlayerCharacter.transaction do
+            @@report << "Updating PC #" + pc.id.to_s + " " + pc.name
 
-        pc.lock!
-        turns_added = ( pc.turns > 160 ? 200 - pc.turns : 40 )
-        pc.turns += turns_added
-        pc.save!
+            pc.lock!
+            turns_added = (pc.turns > 160 ? 200 - pc.turns : 40)
+            pc.turns    += turns_added
+            pc.save!
 
-        @@report << " gained " + turns_added.to_s + " turns."
+            @@report << " gained " + turns_added.to_s + " turns."
 
-        pc.health.lock!
-        #if not alive, no point in damage from disease and checking to see if it killed themz
-        if pc.health.wellness == SpecialCode.get_code('wellness','alive')
-          for infection in pc.illnesses   #take health away at end of day, ratehr than for every turn
-            #future thing to add: see if the pc has th constitution to just immunize to the disease
-            pc.health.HP -= infection.disease.HP_per_turn
-            pc.health.MP -= infection.disease.MP_per_turn
-            @@report << infection.disease.name + " causes " + infection.disease.HP_per_turn.to_s + " HP, " + infection.disease.MP_per_turn.to_s + " MP damage."
+            pc.health.lock!
+            #if not alive, no point in damage from disease and checking to see if it killed themz
+            if pc.health.wellness == SpecialCode.get_code('wellness', 'alive')
+              for infection in pc.illnesses #take health away at end of day, ratehr than for every turn
+                #future thing to add: see if the pc has th constitution to just immunize to the disease
+                pc.health.HP -= infection.disease.HP_per_turn
+                pc.health.MP -= infection.disease.MP_per_turn
+                @@report << infection.disease.name + " causes " + infection.disease.HP_per_turn.to_s + " HP, " + infection.disease.MP_per_turn.to_s + " MP damage."
+              end
+              #see if PC still alive.
+              if pc.health.HP <= 0
+                pc.health.wellness = SpecialCode.get_code('wellness', 'dead')
+                @@report << " Died from untreated disease."
+              end
+            end
+            pc.health.save!
           end
-          #see if PC still alive.
-          if pc.health.HP <= 0
-            pc.health.wellness = SpecialCode.get_code('wellness','dead')
-            @@report << " Died from untreated disease."
-          end
+          @updated_count += 1
+        rescue Exception => e
+          @@report << e
         end
-        pc.health.save!
-      end
-      @updated_count += 1
-      rescue Exception => e
-        @@report << e
       end
     end
     @updated_count
@@ -251,14 +255,14 @@ class Maintenance < ActiveRecord::Base
 
   def self.creature_maintenance
     @@report << "Creature maintenance"
-    Creature.find(:all, :conditions => ['armed = true AND being_fought > 0']).each{|c|
-    begin
-      c.lock!
-      c.number_alive += c.being_fought
-      @@report << "  shifted " + c.being_fought.to_s + " " + c.name + " from being fought to alive; total alive now:" + c.number_alive.to_s
-      c.save!
-    rescue Exception => e
-     @@report << e
-    end }
+    Creature.where(armed: true).where('being_fought > 0').each { |c|
+      begin
+        c.lock!
+        c.number_alive += c.being_fought
+        @@report << "  shifted " + c.being_fought.to_s + " " + c.name + " from being fought to alive; total alive now:" + c.number_alive.to_s
+        c.save!
+      rescue Exception => e
+        @@report << e
+      end }
   end
 end
